refactor: update CheckAlbumMessages signature to use int32 for count
Align count parameter type with TDLib API expectations and use
proper int32 cast to avoid implicit conversions.

+++ test/bdd/steps/04_media_test.go
@@ -49,7 +49,7 @@ func register04MediaSteps(ctx *godog.ScenarioContext, s *scenarioCtx) {
 
 	ctx.Then(`^медиа-альбом появляется во всех целевых чатах в правильном порядке$`, func() error {
 		for _, targetID := range s.env.TargetIDs {
-			msgs, err := s.env.CheckAlbumMessages(targetID, s.prefix, len(testPhotos))
+			msgs, err := s.env.CheckAlbumMessages(targetID, s.prefix, int32(len(testPhotos)))
 			if err != nil {
 				return err
 			}
